Add onImageExtracted callback to image editor on Done
